[feat/#37]: 컴포넌트 분리
